+++ src/Sulu/Bundle/AdminBundle/Resources/public/js/aura_extensions/url-manager.js
@@ -54,7 +54,7 @@
                     }
 
                     return sandbox.template.parse(template, data);
-                }
+                };
             }
         };
     });