+++ src/Sulu/Bundle/ContentBundle/Resources/public/js/components/content/main.js
@@ -1256,7 +1256,7 @@ define([
                     }
 
                     if (!!navigationUrlParams.length) {
-                        navigationUrl += '?' + navigationUrlParams.join('&')
+                        navigationUrl += '?' + navigationUrlParams.join('&');
                     }
 
                     header = {